api/requirements.txt
openai
python-dotenv
fastapi
pydantic
uvicorn[standard]